// client/stocklinx app/src/redux/accessory/constant.ts
export enum accessoryConst {
  FETCH_ACCESSORIES_REQUEST = "FETCH_ACCESSORIES_REQUEST",
  FETCH_ACCESSORIES_SUCCESS = "FETCH_ACCESSORIES_SUCCESS",
  FETCH_ACCESSORIES_FAILURE = "FETCH_ACCESSORIES_FAILURE",
  FETCH_ACCESSORY_REQUEST = "FETCH_ACCESSORY_REQUEST",
  FETCH_ACCESSORY_SUCCESS = "FETCH_ACCESSORY_SUCCESS",
  FETCH_ACCESSORY_FAILURE = "FETCH_ACCESSORY_FAILURE",
  CREATE_ACCESSORY_REQUEST = "CREATE_ACCESSORY_REQUEST",
  CREATE_ACCESSORY_SUCCESS = "CREATE_ACCESSORY_SUCCESS",
  CREATE_ACCESSORY_FAILURE = "CREATE_ACCESSORY_FAILURE",
  UPDATE_ACCESSORY_REQUEST = "UPDATE_ACCESSORY_REQUEST",
  UPDATE_ACCESSORY_SUCCESS = "UPDATE_ACCESSORY_SUCCESS",
  UPDATE_ACCESSORY_FAILURE = "UPDATE_ACCESSORY_FAILURE",
  REMOVE_ACCESSORY_REQUEST = "REMOVE_ACCESSORY_REQUEST",
  REMOVE_ACCESSORY_SUCCESS = "REMOVE_ACCESSORY_SUCCESS",
  REMOVE_ACCESSORY_FAILURE = "REMOVE_ACCESSORY_FAILURE",
  SET_ACCESSORY = "SET_ACCESSORY",
  CLEAR_ACCESSORY = "CLEAR_ACCESSORY",
  SET_ACCESSORIES = "SET_ACCESSORIES",
  CLEAR_ACCESSORIES = "CLEAR_ACCESSORIES",
}